refactor: extract shared handler
Extract unified request handling logic from `handleIngressAPI` into a
shared `handleApiRequest` function.

Related-Task: INTER-2097
