+++ src/handlers/index.ts
@@ -1,3 +1,4 @@
 export { handleStatusPage } from './handleStatusPage'
 export { handleDownloadScript } from './handleDownloadScript'
 export { handleIngressAPI } from './handleIngressAPI'
+export { handleApiRequest } from './handleApiRequest'
